\name{testBeatles-package}
\alias{testBeatles-package}
\alias{testBeatles}
\docType{package}
\title{
\packageTitle{testBeatles}
}
\description{
\packageDescription{testBeatles}
}
\details{
%%  ~~ An overview of how to use the package, ~~
%%  ~~ including the most important functions ~~
}
\author{
\packageAuthor{testBeatles}

Maintainer: \packageMaintainer{testBeatles}
}
\references{
%%  ~~ Literature or other references for background information ~~
}
\keyword{package}

%% Uncomment below to imitate parts of library(help = testBeatles)
%\section{The \file{DESCRIPTION} File}{\packageDESCRIPTION{testBeatles}}
%\section{Documentation Index}{\packageIndices{testBeatles}}

\seealso{
%%  ~~ Optional links to other man pages, e.g. ~~
%%  ~~ \code{\link[<pkg>:<pkg>-package]{<pkg>}} ~~
}
\examples{
%%  ~~ Optional simple examples of the most important functions ~~
}
